add example Zephyr CAN client; get Zephyr CAN subscriptions working

+++ examples/canserver/main.cpp
@@ -100,7 +100,7 @@ int main()
         while (true) {
             sleep(1);
             std::cout << "Publishing report" << std::endl;
-            for (int i = 0; i < moduleRecords.size(); i++) {
+            for (size_t i = 0; i < moduleRecords.size(); i++) {
                 auto increment = 0.25 * (std::rand() % 4);
                 if (std::rand() % 1 == 0) {
                     increment *= -1;
